+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to Sharkable are documented here.
 
-## Unreleased
+## [0.5.3] — 2026-06-30
 
 ### feat